redo 3.1.12 - remove class which does not have to be public
