+++ src/main/java/edu/kis/vh/nursery/list/IntLinkedList.java
@@ -41,4 +41,19 @@ public class IntLinkedList {
         return ret;
     }
 
+    private class Node {
+
+        private final int value;
+        protected Node prev, next;
+
+        public Node(int i) {
+            value = i;
+        }
+
+        public int getValue() {
+            return value;
+        }
+    }
+
+
 }
